Add Accessories & Jewelry System (SM-16) and update index
- Introduced the Accessories & Jewelry System (SM-16) with comprehensive features for additional equipment slots, stat bonuses, and set effects.
- Updated SM-00-index.md to include the new Accessories System in the tier structure and dependencies.
- Enhanced monetization strategies and integration points with existing systems.

+++ projects/adapters/hytale-adapter/src/main/java/com/argonathsystems/adapter/hytaleadapter/converter/PlayerConverter.java
@@ -11,8 +11,8 @@ public class PlayerConverter {
         return new PlayerData(
             player.getUniqueId(),
             player.getName(),
-            player.getHealth(),
-            player.getMaxHealth()
+            (int) player.getHealth(),
+            (int) player.getMaxHealth()
         );
     }
 }
